face_detection/requirements.txt
absl-py==1.4.0
annotated-types==0.5.0
anyio==3.7.1
apache-beam==2.50.0
array-record==0.4.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1694046349000/work
astunparse==1.6.3
avro-python3==1.10.2
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
bbox-visualizer==0.1.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
contextlib2==21.6.0
contourpy==1.1.1
crcmod==1.7
cycler==0.11.0
Cython==3.0.2
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
deep-sort-realtime==1.3.2
deepface==0.0.79
deepsort==0.0.5
dill==0.3.1.1
dm-tree==0.1.8
dnspython==2.4.2
docopt==0.6.2
email-validator==2.0.0.post2
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
etils==1.5.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1692026125334/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fastapi==0.103.2
fastavro==1.8.3
fasteners==0.19
filelock==3.12.4
fire==0.5.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.42.1
fsspec==2023.9.2
gast==0.4.0
gdown==4.7.1
gin-config==0.5.0
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.9.0
hdfs==2.7.2
httpcore==0.18.0
httplib2==0.22.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
immutabledict==3.0.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1693880262622/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1693579759651/work
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1690896916983/work
Jinja2==3.1.2
joblib==1.3.2
jupyter-bbox-widget==0.5.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1695648223557/work
jupyterlab-widgets==3.0.9
kaggle==1.5.16
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
lvis==0.5.3
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mtcnn==0.1.1
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
numpy==1.24.3
nvidia-cublas-cu11==2022.4.8
nvidia-cublas-cu117==11.10.1.25
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu11==8.6.0.163
nvidia-cudnn-cu12==8.9.4.25
oauth2client==4.1.3
oauthlib==3.2.2
object-detection @ file:///home/raichu/Tensorflow/models/research
objsize==0.6.1
opencv-contrib-python==4.8.0.76
opencv-python==4.8.1.78
opt-einsum==3.3.0
orjson==3.9.7
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.1.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==10.0.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1690813113769/work
portalocker==2.8.2
promise==2.3
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
proto-plus==1.22.3
protobuf==3.20.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pydot==1.4.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1691408637400/work
pylabel==0.1.53
pymongo==4.5.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1686601365461/work
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
sacrebleu==2.2.0
scikit-learn==1.1.3
scipy==1.9.3
sentencepiece==0.1.99
seqeval==1.2.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
soupsieve==2.5
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.27.0
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.13.0
tensorflow-hub==0.14.0
tensorflow-io==0.34.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
tensorflow-model-optimization==0.7.5
tensorflow-text==2.13.0
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==2.3.0
text-unidecode==1.3
tf-models-official==2.13.2
tf-slim==1.1.0
threadpoolctl==3.2.0
toml==0.10.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827245914/work
tqdm==4.66.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1694710374095/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1695040754690/work
tzdata==2023.3
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.0
widgetsnbextension==4.0.9
wrapt==1.15.0
zipp==3.17.0
zstandard==0.21.0
